+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This page reports the main releases only and the main changes therein.
 
 ### Changes
 
+* implemented touchpad image navigation (pan, zoom with pinch)
 * added robust path management in retouch area
 * code clanup, remove remnants of obsolete code